requirements.txt
Django==2.2.12
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
django-cors-headers==3.4.0
uWSGI==2.0.19.1
